// dashboard/src/utils/hasura-api/generated/schemas/trackNativeQueryArgs.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { NativeQueryRelationship } from './nativeQueryRelationship';
import type { TrackNativeQueryArgsArguments } from './trackNativeQueryArgsArguments';
import type { TrackNativeQueryArgsType } from './trackNativeQueryArgsType';

export interface TrackNativeQueryArgs {
  source: string;
  root_field_name: string;
  type: TrackNativeQueryArgsType;
  arguments: TrackNativeQueryArgsArguments;
  code: string;
  returns: string;
  comment?: string;
  object_relationships?: NativeQueryRelationship[];
  array_relationships?: NativeQueryRelationship[];
}